+++ scripts/settings.js
@@ -7,7 +7,7 @@ canvas
 
 import { MODULE_ID, MODULES_ACTIVE, SPEED } from "./const.js";
 import { ModuleSettingsAbstract } from "./ModuleSettingsAbstract.js";
-import { PATCHER } from "./patching.js";
+import { log } from "./util.js";
 
 const SETTINGS = {
   CONTROLS: {
@@ -147,8 +147,7 @@ export class Settings extends ModuleSettingsAbstract {
       config: true,
       default: false,
       type: Boolean,
-      requiresReload: false,
-      onChange: value => this.toggleTokenRuler(value)
+      requiresReload: false
     });
 
     register(KEYS.TOKEN_RULER.SPEED_HIGHLIGHTING, {
@@ -158,8 +157,7 @@ export class Settings extends ModuleSettingsAbstract {
       config: true,
       default: false,
       type: Boolean,
-      requiresReload: false,
-      onChange: value => this.toggleSpeedHighlighting(value)
+      requiresReload: false
     });
 
     register(KEYS.TOKEN_RULER.SPEED_PROPERTY, {
@@ -187,8 +185,6 @@ export class Settings extends ModuleSettingsAbstract {
     });
 
     // Initialize the Token Ruler.
-    if ( this.get(KEYS.TOKEN_RULER.ENABLED) ) this.toggleTokenRuler(true);
-    if ( this.get(KEYS.TOKEN_RULER.SPEED_HIGHLIGHTING) ) this.toggleSpeedHighlighting(true);
     this.setSpeedProperty(this.get(KEYS.TOKEN_RULER.SPEED_PROPERTY));
   }
 
@@ -234,18 +230,7 @@ export class Settings extends ModuleSettingsAbstract {
     });
   }
 
-  static toggleTokenRuler(value) {
-    if ( value ) PATCHER.registerGroup("TOKEN_RULER");
-    else PATCHER.deregisterGroup("TOKEN_RULER");
-  }
-
-  static toggleSpeedHighlighting(value) {
-    if ( value ) PATCHER.registerGroup("SPEED_HIGHLIGHTING");
-    else PATCHER.deregisterGroup("SPEED_HIGHLIGHTING");
-  }
-
   static setSpeedProperty(value) { SPEED.ATTRIBUTE = value; }
-
 }
 
 /**
@@ -268,7 +253,7 @@ function toggleTokenRulerWaypoint(context, add = true) {
   const position = canvas.mousePosition;
   const ruler = canvas.controls.ruler;
   if ( !canvas.tokens.active || !ruler || !ruler.active ) return;
-  // console.debug(`${add ? "add" : "remove"}TokenRulerWaypoint`);
+  log(`${add ? "add" : "remove"}TokenRulerWaypoint`);
 
   // Keep track of when we last added/deleted a waypoint.
   const now = Date.now();
@@ -276,7 +261,7 @@ function toggleTokenRulerWaypoint(context, add = true) {
   if ( delta < 100 ) return true; // Throttle keyboard movement once per 100ms
   MOVE_TIME = now;
 
-  // console.debug(`${add ? "adding" : "removing"}TokenRulerWaypoint`);
+  log(`${add ? "adding" : "removing"}TokenRulerWaypoint`);
   if ( add ) ruler._addWaypoint(position);
   else if ( ruler.waypoints.length > 1 ) ruler._removeWaypoint(position); // Removing the last waypoint throws errors.
 }